+++ packages/kiarina-agi-audio/tests/asr_provider_impl/google/_models/test_google_asr_provider.py
@@ -14,7 +14,6 @@ def test_init_and_properties() -> None:
     provider = GoogleASRProvider(settings)
 
     print(str(provider))
-    print(f"client: {provider.client}")
 
 
 @pytest.mark.costly