+++ packages/kiarina-agi-audio/tests/audio_source_impl/file/_models/test_file_audio_source.py
@@ -10,11 +10,11 @@ from kiarina.agi.audio_source_impl.file._models.file_audio_source import (
 )
 
 
-async def test_file_audio_source(speech_audio_file_path: str) -> None:
+async def test_file_audio_source(audio_file_path: str) -> None:
     audio_source = FileAudioSource(FileAudioSourceSettings(start_timestamp=100.0))
     print(f"__str__: {audio_source}")
 
-    async with audio_source.open(speech_audio_file_path):
+    async with audio_source.open(audio_file_path):
         chunks = [chunk async for chunk in audio_source.read()]
 
     assert len(chunks) > 0
@@ -29,7 +29,7 @@ async def test_file_audio_source(speech_audio_file_path: str) -> None:
         )
 
 
-async def test_converts_sample_rate_and_channels(speech_audio_file_path: str) -> None:
+async def test_converts_sample_rate_and_channels(audio_file_path: str) -> None:
     audio_source = FileAudioSource(
         FileAudioSourceSettings(
             chunk_size=256,
@@ -39,7 +39,7 @@ async def test_converts_sample_rate_and_channels(speech_audio_file_path: str) ->
         )
     )
 
-    async with audio_source.open(speech_audio_file_path):
+    async with audio_source.open(audio_file_path):
         chunks = [chunk async for chunk in audio_source.read()]
 
     assert len(chunks) > 0